refactor: simplyfying core-transaction-pool plugin, movin depend impl
