+++ packages/core-transaction-pool-redis/package.json
@@ -24,7 +24,8 @@
     "@arkecosystem/core-transaction-pool": "^0.1.1",
     "@arkecosystem/crypto": "^0.1.1",
     "delay": "^3.0.0",
-    "ioredis": "^3.2.2"
+    "ioredis": "^3.2.2",
+    "lodash": "^4.17.10"
   },
   "publishConfig": {
     "access": "public"
