+++ packages/core-transaction-pool/lib/utils/validation-helpers.js
@@ -3,21 +3,8 @@ const container = require('@arkecosystem/core-container')
 const bs58check = require('bs58check')
 const { configManager } = require('@arkecosystem/crypto')
 const logger = container.resolvePlugin('logger')
-const database = container.resolvePlugin('database')
 
 module.exports = {
-  /**
-   * Gets the blockchain wallet and checks if transaction can be applied - before returning it to forger.
-   * @param {Transaction}
-   * @return {Boolean}
-   */
-  canApplyToBlockchain: transaction => {
-    return database
-      .walletManager
-      .getWalletByPublicKey(transaction.senderPublicKey)
-      .canApply(transaction)
-  },
-
   /**
    * Checks if transaction recipient is on the same network as blockchain
    * @param {Transaction}
